fix videplayer stop
When playing video, the roku will issue two stops every time you stop a video. The recent activity on the server shows two stops at near the exact same millisecond. The reason is the partialresult and fullresult arent the last result. The isScreenClosed is the last. So if you have stops and reporting within the partialresult and fullresult you are being redundant. The isScreenClosed event will handle that. Hope you understand. :)