Update series to allow "jumping" episodes
This adds in nextup capability for the tv series screen. This was already working to jump to season, with these changes it will now correctly jump to episode as well.

My earlier pull which was incorporated already is passing focusedindexitem to the tvscreen it just wasn't being initialized or used properly.