Add Dialog to install beta from within app
This goes one step further than just to announce the new app in a dialog. This also lets the users choose to install the new beta app from within the app with a single press of the OK button. This uses roku ECP control to launch the install process of the new beta directly on roku. This makes it simple for users to install the new app.